refactor: add custom hooks to separate the logic from the visual part of the components
