+++ src/components/BookCard/index.tsx
@@ -1,14 +1,11 @@
 import Link from "next/link";
-import { useRouter } from "next/router";
 
 import * as M from "@mui/material";
 
 import * as S from "./styles";
-import { useAppDispatch } from "@/store/hooks";
-import { addToCart } from "@/store/slices/cartSlice";
-import { CartItem } from "@/types/cart";
+import useBookCard from "./useBookCard";
 
-interface BookCardProps {
+export interface BookCardProps {
   isbn13: string;
   image: string;
   title: string;
@@ -16,31 +13,12 @@ interface BookCardProps {
 }
 
 const BookCard = ({ isbn13, image, title, price }: BookCardProps) => {
-  const router = useRouter();
-  const dispatch = useAppDispatch();
-
-  const titleSlug = title.replace(/,/g, "").replace(/ /g, "-");
-
-  const item: CartItem = {
-    book: {
-      image,
-      isbn13,
-      price,
-      title,
-    },
-    quantity: 1,
-    totalPrice: Number(price.slice(1)),
-  };
-
-  const handleAddToCart = () => {
-    dispatch(addToCart(item));
-  };
-
-  const handleBuy = () => {
-    dispatch(addToCart(item));
-
-    router.push("/cart");
-  };
+  const { titleSlug, handleAddToCart, handleBuy } = useBookCard({
+    image,
+    title,
+    price,
+    isbn13,
+  });
 
   return (
     <M.Card>
