+++ src/components/CartItem/index.tsx
@@ -5,34 +5,20 @@ import AddOutlinedIcon from "@mui/icons-material/AddOutlined";
 
 import * as S from "./styles";
 import { CartItem } from "@/types/cart";
-import { useAppDispatch } from "@/store/hooks";
-import {
-  decreaseQuantity,
-  increaseQuantity,
-  removeFromCart,
-} from "@/store/slices/cartSlice";
 import { formatPrice } from "@/utils/formatting";
+import useCartItem from "./useCartItem";
 
 interface CartItemProps {
   cartItem: CartItem;
 }
 
 const CartItem = ({ cartItem }: CartItemProps) => {
-  const dispatch = useAppDispatch();
-
-  const totalPrice = Number(cartItem.book.price.slice(1)) * cartItem.quantity;
-
-  const handleRemoveFromCart = () => {
-    dispatch(removeFromCart(cartItem.book.isbn13));
-  };
-
-  const handleIncreaseQuantity = () => {
-    dispatch(increaseQuantity(cartItem.book.isbn13));
-  };
-
-  const handleDecreaseQuantity = () => {
-    dispatch(decreaseQuantity(cartItem.book.isbn13));
-  };
+  const {
+    totalPrice,
+    handleRemoveFromCart,
+    handleIncreaseQuantity,
+    handleDecreaseQuantity,
+  } = useCartItem(cartItem);
 
   return (
     <S.Wrapper>
