+++ src/templates/books/search/index.tsx
@@ -1,34 +1,19 @@
-import { useRouter } from "next/router";
-
 import * as M from "@mui/material";
 import SearchOffOutlinedIcon from "@mui/icons-material/SearchOffOutlined";
 
 import * as S from "./styles";
 import BookCard from "@/components/BookCard";
-import { GetServerSideProps } from "next";
-import { api } from "@/services/api";
-import { Book, SearchResultData } from "@/types/books";
-import { ChangeEvent } from "react";
+import { SearchResultData } from "@/types/books";
 import NoContentMessage from "@/components/NoContentMessage";
+import useSearch from "./useSearch";
 
 export interface SearchTemplateProps {
   searchResultData: SearchResultData;
 }
 
 const SearchTemplate = ({ searchResultData }: SearchTemplateProps) => {
-  const router = useRouter();
-  const keyword = router.query.keyword;
-  const totalPages = Math.ceil(Number(searchResultData.total) / 10);
-  const currentPage = router.query.page
-    ? Number(router.query.page as string)
-    : 1;
-
-  const handleChangePage = (event: ChangeEvent<unknown>, page: number) => {
-    router.push({
-      pathname: "/books/search",
-      query: { keyword, page },
-    });
-  };
+  const { handleChangePage, totalPages, currentPage, keyword } =
+    useSearch(searchResultData);
 
   return (
     <S.Wrapper>
